refactor: add tests for BodyTemplate
